broke file serving functions out into seperate file
