+++ lib/nowServerLib.js
@@ -8,35 +8,10 @@ var fs = require('fs');
 var serverScope = {};
 var connectedFuncs = [];
 var disconnectedFuncs = [];
-
-var fileCache = {};
-var nowFileCache;
 var socket;
 
+var fileServer = require('./fileServer.js');
 
-var everyone = {
-  nowScope: {},
-  connected: function(func) {
-    // Instead of using events, we'll just add it to an array of functions that needs to be called
-    if(arguments.length === 0) {
-      for(var i in connectedFuncs) {
-        connectedFuncs[i].apply(this);
-      }
-    } else {
-      connectedFuncs.push(func); 
-    }
-  },
-  disconnected: function(func) {
-    // Instead of using events, we'll just add it to an array of functions that needs to be called
-    if(arguments.length === 0) {
-      for(var i in disconnectedFuncs) {
-        disconnectedFuncs[i].apply(this);
-      }
-    } else {
-      disconnectedFuncs.push(func); 
-    }
-  }
-};
 
 var nowCore = {
   scopes: {},
@@ -102,53 +77,6 @@ var nowCore = {
       }
       scope[data.key] = newVal;
     }
-  },
-  
-  everyoneStore: {
-    set: function(key, val, callback, changes){
-
-      nowUtil.debug("everyoneStore", key + " => " + val);
-      var scopes = [];
-      
-      // Put all relevant scopes in array so we traverse only once
-      for(var i in nowCore.scopes) {
-        scopes.push(nowCore.scopes[i]);
-      }
-      scopes.push(serverScope);
-      
-      nowUtil.mergeChanges(scopes, changes);
-      
-      var data = nowUtil.decycle(val, "now."+key, [function(fqn, func){
-        var multiCaller = nowCore.generateMultiCaller(fqn);
-        nowUtil.createVarAtFqn(fqn, everyone.nowScope, multiCaller);
-        return nowUtil.serializeFunction(fqn, func); 
-      }]);
-      
-      socket.broadcast({type: 'replaceVar', data: {key: key, value: data[0]}});
-      
-      callback();  
-    },
-    remove: function(key){
-      // console.log("remove " + key);
-    }
-  },
-  
-  generateMultiCaller: function(fqn){
-    nowUtil.debug("generateMultiCaller", fqn);
-    return function multicall (){
-      var outputs = {};
-      for(var clientId in nowCore.scopes){
-        nowUtil.debug("Multicaller", "Calling "+fqn+" on client " + clientId);
-        var clientScope = nowCore.scopes[clientId];
-        var theFunction = nowUtil.getVarFromFqn(fqn, clientScope);
-        if(theFunction !== false) {
-          outputs[clientId] = theFunction.apply({now: clientScope}, arguments);
-        } else {
-          nowUtil.debug("Multicaler", "No function found for client " + clientId);
-        }
-      }
-      return outputs;
-    };
   },
 
   handleDisconnection: function(client) {
@@ -218,57 +146,9 @@ var nowCore = {
   }
 };
 
-var serveFile = function(filename, request, response, options){
-  var options = options || {};
 
-  if(fileCache.hasOwnProperty(filename)) {
-    response.writeHead(200);
-    response.write(fileCache[filename]);
-    response.end();
-  } else { 
-    if(filename.indexOf("/now.js") !== -1) {
-      if(nowFileCache === undefined) {
-        fs.readFile(filename, function(err, data){  
-          var host = request.headers.host.split(":");
-          var hostServer = options['host'] || host[0];
-          var hostPort = '80';
-          if(host.length > 1){
-            hostPort = host[1];
-          }
-          hostPort = options['port'] || hostPort;
 
-          var text = data.toString();
-          text = text.replace(/\*\*SERVER\*\*/g, hostServer);
-          text = text.replace(/\*\*PORT\*\*/g, hostPort);
-          
-          fs.readFile(__dirname + "/nowUtil.js", function(err, data){ 
-            var textUtil = data.toString();
-            nowFileCache = {nowUtil: textUtil, now: text};
-            response.writeHead(200, {'content-type': 'text/javascript'});
-            response.write(nowFileCache.nowUtil);
-            response.write(nowFileCache.now);
-            response.end();
-          });
-        });
-      } else {
-        response.writeHead(200, {'content-type': 'text/javascript'});
-        response.write(nowFileCache.nowUtil);
-        response.write(nowFileCache.now);
-        response.end();
-      }
-    } else {
-      fs.readFile(filename, function(err, data){
-        var text = data.toString();
-        response.writeHead(200);
-        response.write(text);
-        response.end();
-        fileCache[filename] = text;
-      });
-    }
-  }
-};
-
-var handleNewConnection = function(client){
+function handleNewConnection(client){
   client.on('message', function(message){
     var messageObj = message;
     if(messageObj !== null && messageObj.hasOwnProperty('type') && nowCore.messageHandlers.hasOwnProperty(messageObj.type)) {
@@ -283,25 +163,7 @@ var handleNewConnection = function(client){
 exports.initialize = function(server, options, socketioOptions){
   var options = options || {};
   // Override the default HTTP server listeners
-  var defaultListeners = server.listeners('request');
-  server.removeAllListeners('request');
-  server.on('request', function(request, response){
-    // Handle only GET requests for /nowjs/* files. Pass all other requests through
-    var i;
-    if(request.method === "GET"){
-      if(request.url.split('?')[0] === "/nowjs/now.js") {
-          serveFile(__dirname + '/now.js', request, response, options);
-      } else {
-        for(i in defaultListeners){
-          defaultListeners[i].call(server, request, response);
-        }
-      }
-    } else {
-      for(i in defaultListeners){
-        defaultListeners[i].call(server, request, response);
-      }
-    }
-  });
+  fileServer.wrapServer(server);
   
   socket = io.listen(server, socketioOptions);
   socket.on('connection', function(client){
@@ -311,8 +173,6 @@ exports.initialize = function(server, options, socketioOptions){
     
   return everyone;
 };
-
-everyone.now = proxy.wrap(nowCore.everyoneStore, everyone.nowScope);
 
 // Handle uncaught exceptions
 process.on('uncaughtException', function (err) {
